Move utility functions into dedicated file
